refactoring 6
