+++ src/request-pipeline/utils.js
@@ -3,6 +3,16 @@ import matchUrl from 'match-url-wildcard';
 import * as headerTransforms from './header-transforms';
 import FileRequest from './file-request';
 import DestinationRequest from './destination-request';
+import * as requestEventInfo from '../session/events/info';
+import promisifyStream from '../utils/promisify-stream';
+import ConfigureResponseEvent from '../session/events/configure-response-event';
+import RequestEvent from '../session/events/request-event';
+import ResponseEvent from '../session/events/response-event';
+import REQUEST_EVENT_NAMES from '../session/events/names';
+import ConfigureResponseEventOptions from '../session/events/configure-response-event-options';
+import { toReadableStream } from '../utils/buffer';
+import { PassThrough } from 'stream';
+import SAME_ORIGIN_CHECK_FAILED_STATUS_CODE from './xhr/same-origin-check-failed-status-code';
 
 export function createReqOpts (ctx) {
     const bodyWithUploads = injectUpload(ctx.req.headers['content-type'], ctx.reqBody);
@@ -90,4 +100,73 @@ export function error (ctx, err) {
         ctx.req.destroy();
     else
         ctx.closeWithError(500, err.toString());
+}
+
+export async function callResponseEventCallbackForProcessedRequest (ctx, configureResponseEvent) {
+    const responseInfo         = requestEventInfo.createResponseInfo(ctx);
+    const preparedResponseInfo = requestEventInfo.prepareEventData(responseInfo, configureResponseEvent.opts);
+    const responseEvent        = new ResponseEvent(configureResponseEvent._requestFilterRule, preparedResponseInfo);
+
+    await ctx.session.callRequestEventCallback(REQUEST_EVENT_NAMES.onResponse, configureResponseEvent._requestFilterRule, responseEvent);
+}
+
+export async function callOnRequestEventCallback (ctx, rule, reqInfo) {
+    const requestEvent = new RequestEvent(ctx, rule, reqInfo);
+
+    await ctx.session.callRequestEventCallback(REQUEST_EVENT_NAMES.onRequest, rule, requestEvent);
+}
+
+export async function callOnResponseEventCallbackForFailedSameOriginCheck (ctx, rule, configureOpts) {
+    const responseInfo = requestEventInfo.createResponseInfo(ctx);
+
+    responseInfo.statusCode = SAME_ORIGIN_CHECK_FAILED_STATUS_CODE;
+
+    const preparedResponseInfo = requestEventInfo.prepareEventData(responseInfo, configureOpts);
+    const responseEvent        = new ResponseEvent(rule, preparedResponseInfo);
+
+    await ctx.session.callRequestEventCallback(REQUEST_EVENT_NAMES.onResponse, rule, responseEvent);
+}
+
+export async function callOnConfigureResponseEventForNonProcessedRequest (ctx) {
+    await ctx.forEachRequestFilterRule(async rule => {
+        const configureResponseEvent = new ConfigureResponseEvent(ctx, rule, ConfigureResponseEventOptions.DEFAULT);
+
+        await ctx.session.callRequestEventCallback(REQUEST_EVENT_NAMES.onConfigureResponse, rule, configureResponseEvent);
+
+        ctx.onResponseEventData.push({ rule, opts: configureResponseEvent.opts });
+    });
+}
+
+export async function callOnResponseEventCallbackWithBodyForNonProcessedRequest (ctx, onResponseEventDataWithBody) {
+    const destResBodyCollectorStream = new PassThrough();
+
+    ctx.destRes.pipe(destResBodyCollectorStream);
+
+    promisifyStream(destResBodyCollectorStream).then(async data => {
+        ctx.saveNonProcessedDestResBody(data);
+
+        const responseInfo = requestEventInfo.createResponseInfo(ctx);
+
+        await Promise.all(onResponseEventDataWithBody.map(async ({ rule, opts }) => {
+            const preparedResponseInfo = requestEventInfo.prepareEventData(responseInfo, opts);
+            const responseEvent        = new ResponseEvent(rule, preparedResponseInfo);
+
+            await ctx.session.callRequestEventCallback(REQUEST_EVENT_NAMES.onResponse, rule, responseEvent);
+        }));
+
+        toReadableStream(data).pipe(ctx.res);
+    });
+}
+
+export async function callOnResponseEventCallbackWithoutBodyForNonProcessedResource (ctx, onResponseEventDataWithoutBody) {
+    const responseInfo = requestEventInfo.createResponseInfo(ctx);
+
+    await Promise.all(onResponseEventDataWithoutBody.map(async item => {
+        const preparedResponseInfo = requestEventInfo.prepareEventData(responseInfo, item.opts);
+        const responseEvent        = new ResponseEvent(item.rule, preparedResponseInfo);
+
+        await ctx.session.callRequestEventCallback(REQUEST_EVENT_NAMES.onResponse, item.rule, responseEvent);
+    }));
+
+    ctx.destRes.pipe(ctx.res);
 }
